e2e: enable the egress authorization suite and delete sdsmint suite
The old sdsmint suite includes three tests: TestSdsmintMintsALeafPerSN,
TestGatewayRefusesANonActorWorkload and TestGatewayRefusesAnUnknownActor.

The egressmitm suite covers the e2e functionality of sdsmint, making
TestSdsmintMintsALeafPerSNI unnecessary.

TestGatewayRefusesANonActorWorkload and TestGatewayRefusesAnUnknownActor
used to live in the sdsmint suite. But their functionality does not
depend on sdsmint. So move them into a suite egressauthz.

New manifest_test.go verifies that the MITM CA is only mounted
on the `sdsmint` container in the `atenet-egress` Deployment.
